refactor(imports): change imports for useGetCollaborators
